--- a/Gemfile
+++ b/Gemfile
@@ -10,7 +10,7 @@ source 'https://rubygems.org'
 #
 # This will help ensure the proper Jekyll version is running.
 # Happy Jekylling!
-gem 'jekyll', '~> 3.9.0'
+gem 'jekyll', '~> 4.2.0'
 
 # This is needed for GitHub Flavored Markdown
 gem 'kramdown-parser-gfm'