--- a/Gemfile
+++ b/Gemfile
@@ -36,6 +36,7 @@ gem 'minima', '~> 2.5'
 # If you have any plugins, put them here!
 group :jekyll_plugins do
   gem 'jekyll-feed', '~> 0.15'
+  gem 'jekyll-include-cache'
   gem 'jekyll-indico', '~> 0.3.0'
   # gem 'jekyll-indico', github: 'iris-hep/jekyll-indico', branch: 'main'
 end