--- a/_config.yml
+++ b/_config.yml
@@ -28,6 +28,7 @@ markdown: kramdown
 theme: minima
 plugins:
   - jekyll-feed
+  - jekyll-include-cache
 
 strict_front_matter: true