--- a/_includes/layout_header_navbar.html
+++ b/_includes/layout_header_navbar.html
@@ -1,7 +1,7 @@
 
 <html lang="{{ site.lang | default: "en-US" }}">
-    {% include head.html %}
+    {% include_cached head.html %}
   <body>
 
-    {% include navbar.html %}
+    {% include_cached navbar.html %}