--- a/_layouts/fellow.html
+++ b/_layouts/fellow.html
@@ -1,6 +1,6 @@
 <!doctype html>
 
-{% include layout_header_navbar.html %}
+{% include_cached layout_header_navbar.html %}
 
 <div id="container">
   <div class="container">
@@ -100,9 +100,9 @@
       </div>
     </main>
   </div>
-  {% include footer.html %}
+  {% include_cached footer.html %}
 </div>
 
-{% include layout_analytics.html %}
+{% include_cached layout_analytics.html %}
 </body>
 </html>