--- a/_layouts/focus-area.html
+++ b/_layouts/focus-area.html
@@ -1,6 +1,6 @@
 <!doctype html>
 
-{% include layout_header_navbar.html %}
+{% include_cached layout_header_navbar.html %}
 
     <div id="container">
       <div class="container">
@@ -56,7 +56,7 @@
 {% endif %}
 
 
-{% include get_pres_list.html %}
+{% assign sorted_presentations = site.data['sorted_presentations'] %}
 {% assign page_focus_area = page.name | remove: ".md" %}
 
 {% assign pres_list = "" | split: "," %}
@@ -93,7 +93,7 @@
 </section>
 {% endif %}
 
-{% include get_pub_list.html %}
+{% assign sorted_publications = site.data['sorted_publications'] %}
 
 {% assign pubs = "" | split: "," %}
 {% for pub in sorted_publications %}
@@ -165,10 +165,10 @@
 
         </main>
       </div>
-      {% include footer.html %}
+      {% include_cached footer.html %}
     </div>
 
-  {% include layout_analytics.html %}
+  {% include_cached layout_analytics.html %}
   </body>
 </html>