--- a/_layouts/main.html
+++ b/_layouts/main.html
@@ -12,7 +12,7 @@
           {{ content }}
         </main>
       </div>
-      {% include footer.html %}
+      {% include_cached footer.html %}
     </div>
 
   {% if site.google_analytics %}