--- a/_layouts/presentations.html
+++ b/_layouts/presentations.html
@@ -1,6 +1,6 @@
 <!doctype html>
 
-{% include layout_header_navbar.html %}
+{% include_cached layout_header_navbar.html %}
 
   <div id="container">
       <div class="container">
@@ -25,10 +25,10 @@
           {{ content }}
         </main>
       </div>
-      {% include footer.html %}
+      {% include_cached footer.html %}
     </div>
 
-  {% include layout_analytics.html %}
+  {% include_cached layout_analytics.html %}
 
   </body>
 </html>