--- a/_plugins/raise_error.rb
+++ b/_plugins/raise_error.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Jekyll
+  # This will add a filter to raise an exception
+  module ExceptionFilter
+    def raise_error(msg)
+      bad_file = @context.registers[:page]['path']
+      err_msg = "On #{bad_file}: #{msg}"
+      raise err_msg
+    end
+  end
+end
+
+Liquid::Template.register_filter(Jekyll::ExceptionFilter)