--- a/pages/about/advisory-panel.md
+++ b/pages/about/advisory-panel.md
@@ -13,7 +13,7 @@ management on the Institute’s overall goals and evolving plans.**
   <div class="row">
   {% for member in site.data.orgs.advisory-panel.personnel  %}
      {% assign person = site.data.people[member] %}
-     {% include standard_person_card.md %}
+     {% include standard_person_card.md person=person %}
   {% endfor %}
   </div>
 </div>