--- a/pages/presentations/byinstitution.md
+++ b/pages/presentations/byinstitution.md
@@ -4,7 +4,7 @@ layout: presentations
 title: Presentations by Institution
 ---
 
-{% include get_pres_list.html %}
+{% assign sorted_presentations = site.data['sorted_presentations'] %}
 
 <!--
   0     1       2      3       4          5           6          7            8