This updates the versions of dependencies; notably Jekyll 4.2 is now used. Adds caching and moves some expensive tasks to Ruby, reducing the generation time from 40 seconds to about 6-7 seconds (once remote information has been cached, of course). Should also be faster when using Docker, too. Based on the output from `bundle exec jekyll build --profile`. Using caching in a few places; we should eventually update our plugins to use Jekyll 4 caching.


Before:
```
Filename                                       | Count |    Bytes |   Time
-----------------------------------------------+-------+----------+-------
_includes/get_pres_list.html                   |    45 |    0.22K | 24.277
_layouts/project.html                          |    35 |  749.73K | 23.790
_includes/list_project_presentations.html      |    35 |   56.86K | 18.958
_layouts/focus-area.html                       |     7 |  358.37K |  5.028
_includes/layout_header_navbar.html            |   139 | 1498.85K |  3.758
_includes/get_pub_list.html                    |    44 |    0.17K |  3.667
_includes/head.html                            |   140 |  801.61K |  3.363
_includes/list_project_publications.html       |    35 |   31.06K |  2.973
_layouts/default.html                          |    55 | 1304.68K |  1.502
_layouts/fellow.html                           |    35 |  771.29K |  1.228
pages/presentations/byinstitution.md           |     1 |  246.65K |  0.712
pages/presentations/byarea.md                  |     1 |  224.47K |  0.662
pages/presentations/bymonth.md                 |     1 |  170.49K |  0.608
_includes/list_videos.html                     |    43 |  164.28K |  0.532
_includes/list_project_team.html               |    35 |    8.16K |  0.504
_includes/navbar.html                          |   140 |  701.23K |  0.380
_includes/get_all_videos.html                  |    43 |    0.04K |  0.318
_includes/print_pres.html                      |    33 |  833.14K |  0.210
_layouts/presentations.html                    |     5 |  918.83K |  0.139
pages/events.md                                |     1 |   14.76K |  0.103
pages/docs/nsfpar.md                           |     1 |   42.77K |  0.102
pages/publications/all.md                      |     1 |   31.86K |  0.094
pages/fellow_projects.md                       |     1 |   47.45K |  0.093
pages/videos.html                              |     1 |   16.12K |  0.077
...

                    done in 39.623 seconds.
```


After:

```
Build Process Summary:

| PHASE      |   TIME |
+------------+--------+
| RESET      | 0.0001 |
| READ       | 0.4617 |
| GENERATE   | 2.6724 |
| RENDER     | 3.2272 |
| CLEANUP    | 0.0194 |
| WRITE      | 0.4271 |
+------------+--------+
| TOTAL TIME | 6.8079 |


Site Render Stats:

| Filename                                            | Count |    Bytes |  Time |
+-----------------------------------------------------+-------+----------+-------+
| _layouts/project.html                               |    35 |  744.72K | 1.139 |
| _includes/list_project_team.html                    |    35 |    8.16K | 0.522 |
| _includes/list_videos.html                          |    43 |  164.28K | 0.505 |
| _layouts/focus-area.html                            |     7 |  359.04K | 0.431 |
| _includes/get_all_videos.html                       |    43 |    0.04K | 0.331 |
| feed.xml                                            |     1 |  141.21K | 0.256 |
| _layouts/fellow.html                                |    35 |  772.37K | 0.233 |
| _includes/print_pres.html                           |  3177 |  833.14K | 0.204 |
| _includes/list_project_presentations.html           |    35 |   56.69K | 0.191 |
| pages/presentations/byinstitution.md                |     1 |  246.65K | 0.164 |
| pages/events.md                                     |     1 |   14.76K | 0.108 |
| pages/fellow_projects.md                            |     1 |   47.45K | 0.096 |
| pages/presentations/byarea.md                       |     1 |  224.46K | 0.087 |
| _includes/list_project_publications.html            |    35 |   30.92K | 0.087 |
| pages/videos.html                                   |     1 |   16.12K | 0.075 |
...
+-----------------------------------------------------+-------+----------+-------+
| TOTAL (for 50 files)                                |  4328 | 5901.16K | 4.996 |

                    done in 6.827 seconds.
```